Update test and move out of sidekiq dir
